[Refactor] Split troubleshooting IPC out to handler
